migration

+++ database/migrations/2019_10_24_165037_create_blog_categories_table.php
@@ -14,8 +14,8 @@ class CreateBlogCategoriesTable extends Migration
     public function up()
     {
         Schema::create('blog_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned()->default(1);
+            $table->bigIncrements('id');
+            $table->bigInteger('parent_id')->unsigned()->default(1);
 
             $table->string('slug')->unique();
             $table->string('title');
